Removed sample types
